bhw1-payments/requirements.txt
rpaframework==7.6.0
requests==2.25.1
google-api-python-client==2.1.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
pandas==1.2.2
PyDrive==1.3.1
retry==0.9.2
ta-bitwarden-cli
